backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
pydantic[dotenv]==2.5.2
pydantic-settings